+++ providers/src/airflow/providers/mongo/CHANGELOG.rst
@@ -27,6 +27,19 @@
 Changelog
 ---------
 
+4.2.2
+.....
+
+Bug Fixes
+~~~~~~~~~
+
+* ``fix(providers/mongo): prevent applying lower method on boolean field (#43024)``
+
+
+.. Below changes are excluded from the changelog. Move them to
+   appropriate section above if needed. Do not delete the lines(!):
+   * ``Split providers out of the main "airflow/" tree into a UV workspace project (#42505)``
+
 4.2.1
 .....